Move utils adreno functinos to separate shared file
